fix z3varcontext bug
turns out that the order of the z3 model had an effect on the order
that we union things in the egraph. and i guess unioning things in the
egraph in different orders changes the result. I think this has to do
something with the fact that egg doesn't maintain the egraph
invariants unless you do a `.rebuild()` explicitly. changing it to
rebuild after unioning things fixes the egraph depending on the order
of the z3 model.