instantiation minifier for debugging
algorithm:
starts with a bag of instantiations. the goal is to find the minimal set
that allows us to progress to the desired depth.

we will remove one at a time.
if removing it:
succeeds? -> save to remove, continue with next instantiation
fails? -> this instantiation was necessary (at least with the current
                                            bag of instantiations)

We do this test on each instantiation, removing unnecessary ones
we keep going through the list until we stop removing instantiations